+++ src/core4inet/linklayer/inControl/AS6802/CTInControl.h
@@ -170,11 +170,11 @@ void CTInControl<IC>::handleParameterChange(const char* parname)
 
     if (!parname || !strcmp(parname, "ct_mask"))
     {
-        this->ctMask = static_cast<uint32_t>(cComponent::par("ct_mask").longValue());
+        this->ctMask = cComponent::par("ct_mask");
     }
     if (!parname || !strcmp(parname, "ct_marker"))
     {
-        this->ctMarker = static_cast<uint32_t>(cComponent::par("ct_marker").longValue());
+        this->ctMarker = cComponent::par("ct_marker");
     }
     if (!parname || !strcmp(parname, "ct_incomings"))
     {
